requirements.txt
websocket-client
python-dotenv
python-dotenv
